knotx-adapters - move source to new modules
