+++ knotx-adapters/knotx-adapter-action-http/pom.xml
@@ -1,4 +1,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Knot.x - Reactive microservice assembler - HTTP Action Adapter
+
+    Copyright (C) 2016 Cognifide Limited
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+         http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
@@ -10,11 +29,11 @@
   </parent>
 
   <artifactId>knotx-adapter-action-http</artifactId>
-  <name>Knot.x - Reactive microservice assembler - Adapters API</name>
+  <name>Knot.x - Reactive microservice assembler - HTTP Action Adapter</name>
 
   <properties>
     <!-- the main verticle class name -->
-    <main.verticle>@@@@FIXME@@@</main.verticle>
+    <main.verticle>com.cognifide.knotx.adapter.action.http.HttpActionAdapterVerticle</main.verticle>
   </properties>
 
   <dependencies>
