+++ knotx-adapters/knotx-adapter-action-http/src/test/resources/logback-test.xml
