+++ knotx-adapters/knotx-adapter-api/src/test/java/com/cognifide/knotx/adapter/api/placeholders/UriTransformerPlaceholderTest.java
@@ -1,0 +1,33 @@
+/*
+ * Knot.x - Reactive microservice assembler - Adapter API
+ *
+ * Copyright (C) 2016 Cognifide Limited
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.cognifide.knotx.adapter.api.placeholders;
+
+import java.util.List;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class UriTransformerPlaceholderTest {
+
+  @Test
+  public void getPlaceholders_whenGivenUrlWithPlaceholders_expectPlaceholdersExtractedInArray() {
+    List<String> placeholders =
+        UriTransformer.getPlaceholders("/dssds/{first.aa}/dsu/{second}");
+    Assert.assertArrayEquals(placeholders.toArray(), new String[]{"first.aa", "second"});
+  }
+}
