+++ knotx-adapters/knotx-adapter-service-http/src/test/resources/first-response.json
