+++ knotx-adapters/knotx-adapter-service-http/src/test/resources/knotx-service-adapter-http-test.json
@@ -1,0 +1,36 @@
+{
+  "verticles": {
+    "com.cognifide.knotx.adapter.service.http.HttpServiceAdapterVerticle": {
+      "config": {
+        "address": "knotx.adapter.service.http",
+        "client.options": {
+          "maxPoolSize": 1000,
+          "keepAlive": false
+        },
+        "services": [
+          {
+            "path": "/service/mock/.*",
+            "domain": "localhost",
+            "port": 3000,
+            "allowed.request.headers": [
+              "Content-Type",
+              "X-*"
+            ]
+          },
+          {
+            "path": "/service/.*",
+            "domain": "localhost",
+            "port": 8080
+          }
+        ]
+      }
+    },
+    "com.cognifide.knotx.mocks.MockServiceVerticle": {
+      "config": {
+        "mock.data.root": "mock/service",
+        "bouncing" : true,
+        "http.port": 3000
+      }
+    }
+  }
+}
