--- knotx-core/knotx-service-adapter-http/src/main/resources/knotx-service-adapter-http.json
@@ -1,27 +1,0 @@
-{
-  "address": "knotx.core.http.service-adapter",
-  "client.options": {
-    "maxPoolSize": 1000,
-    "keepAlive": false
-  },
-  "services": [
-    {
-      "path": "/service/mock/.*",
-      "domain": "localhost",
-      "port": 3000,
-      "allowed.request.headers": [
-        "Content-Type",
-        "X-*"
-      ]
-    },
-    {
-      "path": "/service/.*",
-      "domain": "localhost",
-      "port": 8080,
-      "allowed.request.headers": [
-        "Content-Type",
-        "X-*"
-      ]
-    }
-  ]
-}
