--- knotx-core/knotx-service-adapter-http/src/test/resources/knotx-service-adapter-http-test.json
@@ -1,36 +1,0 @@
-{
-  "verticles": {
-    "com.cognifide.knotx.core.serviceadapter.http.HttpServiceAdapterVerticle": {
-      "config": {
-        "address": "knotx.core.http.service-adapter",
-        "client.options": {
-          "maxPoolSize": 1000,
-          "keepAlive": false
-        },
-        "services": [
-          {
-            "path": "/service/mock/.*",
-            "domain": "localhost",
-            "port": 3000,
-            "allowed.request.headers": [
-              "Content-Type",
-              "X-*"
-            ]
-          },
-          {
-            "path": "/service/.*",
-            "domain": "localhost",
-            "port": 8080
-          }
-        ]
-      }
-    },
-    "com.cognifide.knotx.mocks.MockServiceVerticle": {
-      "config": {
-        "mock.data.root": "mock/service",
-        "bouncing" : true,
-        "http.port": 3000
-      }
-    }
-  }
-}
